Adressing Luke's review comments

+++ Paranormal/Paranormal/Panels/Preview/PreviewSpriteUtils.swift
@@ -3,14 +3,14 @@ import AppKit
 
 class PreviewSpriteUtils: NSObject{
 
-    class func resizeSprite(sprite: CCSprite, toWidth: Float, toHeight: Float) {
-        sprite.scaleX = toWidth / Float(sprite.contentSize.width)
-        sprite.scaleY = toHeight / Float(sprite.contentSize.height)
+    class func resizeSprite(sprite: CCSprite, toWidth: CGFloat, toHeight: CGFloat) {
+        sprite.scaleX = Float(toWidth) / Float(sprite.contentSize.width)
+        sprite.scaleY = Float(toHeight) / Float(sprite.contentSize.height)
     }
 
-    class func resizeSpriteWithoutWarp(sprite: CCSprite, toWidth: Float, toHeight: Float) {
-        let scale = min(toWidth / Float(sprite.contentSize.width),
-            toHeight / Float(sprite.contentSize.height))
+    class func resizeSpriteWithoutWarp(sprite: CCSprite, toWidth: CGFloat, toHeight: CGFloat) {
+        let scale = min(Float(toWidth) / Float(sprite.contentSize.width),
+            Float(toHeight) / Float(sprite.contentSize.height))
         sprite.scaleX = scale
         sprite.scaleY = scale
     }
@@ -28,5 +28,4 @@ class PreviewSpriteUtils: NSObject{
             rotated: false, offset: CGPointMake(0.0, 0.0), originalSize: image.size)
         return spriteFrame
     }
-
 }
